+++ nuttx/configs/cloudctrl/src/cloudctrl-internal.h
@@ -47,7 +47,7 @@
 #include <stdint.h>
 
 /****************************************************************************************************
- * Definitions
+ * Pre-processor Definitions
  ****************************************************************************************************/
 /* Configuration ************************************************************************************/
 /* How many SPI modules does this chip support? */
