+++ nuttx/configs/cloudctrl/src/stm32_adc.c
@@ -1,6 +1,5 @@
 /************************************************************************************
- * configs/cloudctrl/src/up_adc.c
- * arch/arm/src/board/up_adc.c
+ * configs/cloudctrl/src/stm32_adc.c
  *
  *   Copyright (C) 2011-2012 Gregory Nutt. All rights reserved.
  *   Author: Gregory Nutt <gnutt@nuttx.org>
@@ -56,7 +55,7 @@
 #ifdef CONFIG_ADC
 
 /************************************************************************************
- * Definitions
+ * Pre-processor Definitions
  ************************************************************************************/
 /* Configuration ********************************************************************/
 /* Up to 3 ADC interfaces are supported */
