+++ nuttx/configs/cloudctrl/src/stm32_cxxinitialize.c
@@ -1,6 +1,5 @@
 /************************************************************************************
- * configs/cloudctrl/src/up_cxxinitialize.c
- * arch/arm/src/board/up_cxxinitialize.c
+ * configs/cloudctrl/src/stm32_cxxinitialize.c
  *
  *   Copyright (C) 2012 Gregory Nutt. All rights reserved.
  *   Author: Gregory Nutt <gnutt@nuttx.org>
@@ -51,7 +50,7 @@
 #if defined(CONFIG_HAVE_CXX) && defined(CONFIG_HAVE_CXXINITIALIZE)
 
 /************************************************************************************
- * Definitions
+ * Pre-processor Definitions
  ************************************************************************************/
 /* Debug ****************************************************************************/
 /* Non-standard debug that may be enabled just for testing the static constructors */
