+++ nuttx/configs/cloudctrl/src/stm32_nsh.c
@@ -1,6 +1,5 @@
 /****************************************************************************
- * config/cloudctrl/src/up_nsh.c
- * arch/arm/src/board/up_nsh.c
+ * config/cloudctrl/src/stm32_nsh.c
  *
  *   Copyright (C) 2012 Gregory Nutt. All rights reserved.
  *   Author: Gregory Nutt <gnutt@nuttx.org>
@@ -50,7 +49,7 @@
 #include "cloudctrl-internal.h"
 
 /****************************************************************************
- * Pre-Processor Definitions
+ * Pre-processor Definitions
  ****************************************************************************/
 /* Configuration ************************************************************/
